+++ config/importmap.rb
@@ -1,1 +1,2 @@
+pin "admin_suite_application", to: "admin_suite_application.js"
 pin_all_from File.expand_path("../app/javascript/controllers", __dir__), under: "controllers"